StudyFlow/requirements.txt
streamlit
google-generativeai
faiss-cpu
numpy
PyPDF2
python-dotenv
langchain
langchain-community
langchain-google-genai
fpdf
markdown
requests
